Add DeltaRangeAdapter class
This class models a DeltaRange measurement. It does computations by
using two legs, called referenceLeg and otherLeg, which are
GNRangeAdapters. The measurement timetagging is at reception time
of the referenceLeg. The range computation is first done for the
referenceLeg, and then the spacecraft transmission time for the
referenceLeg is taken as transmission time for the otherLeg, and
from that the range computation for the otherLeg is done. Afterwards,
the two ranges for the legs are subtracted.

This introduces a new semantic with ErrorModels. The error model is
assigned per baseline, rather than per groundstation. To accomplish this,
in the GMAT script an ErrorModel of Type 'DeltaRange' must be created,
and this ErrorModel must be assigned to the two Groundstations
participating in the baseline.